Some minor tweaks to the JDK recipe.
It’s still ugly, but you can use the JDK extension to detect if the JDK
is installed. Oracle installs each JDK as it’s own little bundle, so
there isn’t a way to detect the newest one without needing to write a
specific extension attribute XML for it, which I don’t want to do
because I like my templates.