CURLRedirectDownloader.py
While I was working on my Soundslides recipe I discovered that
providing a URL without a filename at the end made the included CURL
downloader fail. I I modified it so it does one CURL loop to determine
the file location at the end of a string of redirects, and then a
second curl loop to download the file.