Adobe Digital Editions flatpkg modifcations
I discovered that the pkg from adobe for digital editions was firing
off the application once the installer finished, which was causing
problems during imaging. I wrote a little python (my first original
script in python! :D) which modifies the postinstall script so that
doesn’t happen anymore. Though the installer is no longer signed by
Adobe, I think that’s a small price to pay (especially since it was
already codesignatured at the start of the policy).