OneDrive recipe complete.
My OneDrive recipe contains some configuration scripts that allow quick
provisioning of OneDrive for Business through Overrides. Out of the
box, the configuration portions of the scripts are disabled, so the one
drive scripts folder will need to be copied to your overrides folder,
and the SCRIPTS variable will need to be updated to point to the
override scripts folder.