Re-add logic folder for GarageBand
So it turns out that some loops are stored in the Logic folder. I
advise that everyone clear out their App Support Garageband and Logic
folders and delete the Apple Loops folder under Audio and re-install
the basic set of sounds for Garageband (it prompts you when you starts
it after delete those folders).